refact: utils폴더 ranking & timer 관련 파일 통합
